+++ lib.h
@@ -1,11 +1,11 @@
 #pragma once
-#include <stdint.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif // __cplusplus
-void reverse(char *s, uint64_t len);
-void upper_case_first(char *s, uint64_t len);
+void reverse(char *s, size_t len);
+void upper_case_first(char *s, size_t len);
 
 #ifdef __cplusplus
 }
